RavenDB-20652 Removing all legacy comparers
