+++ src/Corax/Queries/BoostingMatch.cs
@@ -4,7 +4,7 @@ using System.Diagnostics;
 using System.Numerics;
 using System.Runtime.CompilerServices;
 using Corax.Mappings;
-using Corax.Queries.SortingMatches.Comparers;
+using Corax.Queries.SortingMatches;
 
 namespace Corax.Queries
 {
