+++ src/Corax/Queries/SortingMatches/IMatchComparer.cs
@@ -5,7 +5,7 @@ using Corax.Mappings;
 using Corax.Utils.Spatial;
 using Spatial4n.Shapes;
 
-namespace Corax.Queries.SortingMatches.Comparers
+namespace Corax.Queries.SortingMatches
 {
     public enum MatchCompareFieldType : ushort
     {
@@ -35,29 +35,5 @@ namespace Corax.Queries.SortingMatches.Comparers
         SpatialUnits Units { get; }
         
         IPoint Point { get; }
-    }
-    
-    internal static class BasicComparers
-    {
-        [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        internal static int CompareAscending(ReadOnlySpan<byte> x, ReadOnlySpan<byte> y)
-        {
-            return x.SequenceCompareTo(y);
-        }
-
-        [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        internal static int CompareAscending<T>(T x, T y)
-        {
-            if (typeof(T) == typeof(long))
-            {
-                return Math.Sign((long)(object)y - (long)(object)x);
-            }
-            else if (typeof(T) == typeof(double))
-            {
-                return Math.Sign((double)(object)y - (double)(object)x);
-            }
-
-            throw new NotSupportedException("Not supported");
-        }
     }
 }
