+++ src/Corax/Utils/OrderMetadata.cs
@@ -1,8 +1,6 @@
-﻿using System;
-using System.Runtime.CompilerServices;
+﻿using System.Runtime.CompilerServices;
 using Corax.Mappings;
-using Corax.Queries;
-using Corax.Queries.SortingMatches.Comparers;
+using Corax.Queries.SortingMatches;
 using Corax.Utils.Spatial;
 using Spatial4n.Shapes;
