+++ src/Raven.Server/Documents/Queries/QueryBuilderHelper.cs
@@ -5,7 +5,7 @@ using System.Runtime.CompilerServices;
 using Corax;
 using Corax.Mappings;
 using Corax.Queries;
-using Corax.Queries.SortingMatches.Comparers;
+using Corax.Queries.SortingMatches;
 using Corax.Utils;
 using Raven.Client.Documents.Indexes;
 using Raven.Client.Documents.Indexes.Spatial;
